+++ ApogeeVGC/Sim/Actions/TeamAction.cs
@@ -22,7 +22,7 @@ public record TeamAction : IAction
     }
 
     public double Priority { get; init; }
-    public int Speed => 1;
+    public double Speed => 1;
     public required Pokemon Pokemon { get; init; }
     public int Index { get; init; }
