+++ ApogeeVGC/Sim/BattleClasses/IPriorityComparison.cs
@@ -6,7 +6,7 @@ public interface IPriorityComparison
 {
     IntFalseUnion Order { get; }
     double Priority { get; }
-    int Speed { get; }
+    double Speed { get; }
     int SubOrder { get; }
     int EffectOrder { get; }
     //EffectHolder? EffectHolder { get; }
